pyproject.toml
[tool.poetry]
name = "duo-cli"
version = "0.1.0"
license = "MIT"
readme = "README.md"
description = "Duo Mobile CLI for generating TOTP codes."
keywords = ["totp", "duo", "otp", "cli"]
authors = ["Nathan Cahill <nathan@nathancahill.com>"]
homepage = "https://github.com/nathancahill/duo-cli"
repository = "https://github.com/nathancahill/duo-cli"
include = [
    "LICENSE",
]
classifiers = [
    "Environment :: Console",
]

[tool.poetry.dependencies]
python = "^3.6"
PyInquirer = "^1.0.3"
passlib = "^1.7.4"
PyNaCl = "^1.4.0"
click = "^7.1.2"
pyperclip = "^1.8.2"
rich = "^10.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
duo = "duo_cli.cli:main"
